requirements.txt
Django
django-userena
django-crispy-forms

